+++ src/cxx/wrapper.h
@@ -48,4 +48,4 @@ protected:
 	float _gain;
 };
 
-TFruityPlug& create_plug_instance_c(TFruityPlugHost& Host, int64_t Tag);
+TFruityPlug& create_plug_instance_c(TFruityPlugHost& Host, int Tag);
